# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BioMarket API",
    "description": "Documentacion OpenAPI para productos, inventario y pedidos de BioMarket.",
    "contact": { "name": "BioMarket", "email": "admin@biomarket.com" },
    "license": { "name": "Proyecto academico" },
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Servidor local" }
  ],
  "tags": [
    {
      "name": "Pedidos",
      "description": "Consultas REST de pedidos y actualizacion de estados"
    },
    {
      "name": "Productos",
      "description": "Operaciones REST para el CRUD de productos e inventario"
    }
  ],
  "paths": {
    "/api/products/{id}": {
      "get": {
        "tags": ["Productos"],
        "summary": "Consultar producto",
        "description": "Obtiene un producto por su identificador.",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Productos"],
        "summary": "Actualizar producto",
        "description": "Actualiza los datos principales de un producto.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Productos"],
        "summary": "Eliminar producto",
        "description": "Elimina un producto por su identificador.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": { "204": { "description": "No Content" } }
      }
    },
    "/api/products": {
      "get": {
        "tags": ["Productos"],
        "summary": "Listar productos",
        "description": "Lista productos con filtros opcionales por nombre, categoria y orden.",
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "productName",
            "in": "query",
            "description": "Texto para buscar por nombre del producto",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "Identificador de la categoria",
            "required": false,
            "schema": { "type": "integer", "format": "int32" }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Criterio de orden: name o category",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ProductResponse" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Productos"],
        "summary": "Crear producto",
        "description": "Registra un producto nuevo desde datos JSON.",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ProductRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/ProductResponse" }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}/status": {
      "patch": {
        "tags": ["Pedidos"],
        "summary": "Actualizar estado del pedido",
        "description": "Cambia el estado administrativo de un pedido.",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderStatusRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/OrderResponse" }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Listar pedidos",
        "description": "Lista pedidos registrados con filtros opcionales por estado, email y rango de fechas.",
        "operationId": "listOrders",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Estado exacto del pedido",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "email",
            "in": "query",
            "description": "Texto contenido en el email del cliente",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "fromDate",
            "in": "query",
            "description": "Fecha inicial del filtro",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          },
          {
            "name": "toDate",
            "in": "query",
            "description": "Fecha final del filtro",
            "required": false,
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/OrderResponse" }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Consultar pedido",
        "description": "Obtiene un pedido con cliente y detalles.",
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int32" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/OrderResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProductRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "description": { "type": "string", "minLength": 1 },
          "price": { "type": "number", "minimum": 0.01 },
          "stock": { "type": "integer", "format": "int32", "minimum": 0 },
          "image": { "type": "string" },
          "categoryId": { "type": "integer", "format": "int32" }
        },
        "required": ["categoryId", "description", "name", "price", "stock"]
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "idProduct": { "type": "integer", "format": "int32" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer", "format": "int32" },
          "image": { "type": "string" },
          "categoryId": { "type": "integer", "format": "int32" },
          "categoryName": { "type": "string" }
        }
      },
      "OrderStatusRequest": {
        "type": "object",
        "properties": { "status": { "type": "string", "minLength": 1 } },
        "required": ["status"]
      },
      "ClientSummary": {
        "type": "object",
        "properties": {
          "idClient": { "type": "integer", "format": "int32" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" }
        }
      },
      "OrderDetailSummary": {
        "type": "object",
        "properties": {
          "idDetail": { "type": "integer", "format": "int32" },
          "productId": { "type": "integer", "format": "int32" },
          "productName": { "type": "string" },
          "quantity": { "type": "integer", "format": "int32" },
          "subtotal": { "type": "number" }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "idOrder": { "type": "integer", "format": "int32" },
          "orderDate": { "type": "string", "format": "date-time" },
          "total": { "type": "number" },
          "status": { "type": "string" },
          "client": { "$ref": "#/components/schemas/ClientSummary" },
          "details": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderDetailSummary" }
          }
        }
      }
    }
  }
}